Add directories for each weapon icon type
This is the start of a rework of the weapon icon directory.
Instead of grouping by creator, the icons shall be grouped
by weapon type instead. As a first step, the icons in the
Artist Unknown directory has been moved.
All icons will be prefaced with the creator,
and in some cases the context.